Do not show footer in entry.html if not needed.
If there is no site.read_time or entry.date set, do not show an empty
footer as this only adds some whitespace. In a collection without
excerpts the whitespace between headers is already quite a lot due to
the article tag, so at least, do not add additional unnecessary
whitespace.

Signed-off-by: Vangelis Tasoulas <vangelis@tasoulas.net>