Clarified where to customize scss variables
The documentation stated that lines to change variable names should be added to main.scss after @import statements. I tried this and it did not work for me. 

I then added these lines after the import statements and the customization works perfectly. 

According to [this post](https://stackoverflow.com/questions/17089717/how-to-overwrite-scss-variables-when-compiling-to-css) default variables (such as $accent-color) can be overwritten ahead of time. But changing the value of a variable after it has been used will not have any effect on the earlier parts of the CSS where the variable was used. 

Therefore, it seems like variables have to be overwritten after the import statements, and there is no way to overwrite variables that were not declared with !default.